added _minimum_image_ortho_lazy() to calc_distances.h
This function assumes that particles are already packed into the box.
For O(n^2) algorithms such as the distance_array() functions, it is faster
to first pack all coordinates into the box (that's only O(n)) and then use
the new and faster lazy minimum image routine in the O(n^2) part.
And, yes, that's what has been done.